# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car API",
        "description": "API для управления автомобилями, комплектациями, опциями и ценами",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/prices": {
            "get": {
                "tags": [
                    "Prices"
                ],
                "summary": "Получить список всех цен",
                "operationId": "63348bf88df6d4eb68dfd5192429af28",
                "responses": {
                    "200": {
                        "description": "Список цен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "configuration_id": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 35000
                                            },
                                            "start_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-04-08"
                                            },
                                            "end_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-12-31"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Prices"
                ],
                "summary": "Создать новую цену",
                "operationId": "6a8846778186567e81a8ae51b36fabbd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "configuration_id",
                                    "price",
                                    "start_date"
                                ],
                                "properties": {
                                    "configuration_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 35000
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-08"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Цена создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "configuration_id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 35000
                                        },
                                        "start_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-04-08"
                                        },
                                        "end_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-12-31"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/prices/{id}": {
            "get": {
                "tags": [
                    "Prices"
                ],
                "summary": "Получить цену по ID",
                "operationId": "a573c52428ed8649d28f25cf8388580b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID цены",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Цена найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "configuration_id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 35000
                                        },
                                        "start_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-04-08"
                                        },
                                        "end_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-12-31"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Цена не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Prices"
                ],
                "summary": "Обновить цену",
                "operationId": "3feb2ee9c12dde0605750f30ae506338",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID цены",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "configuration_id",
                                    "price",
                                    "start_date"
                                ],
                                "properties": {
                                    "configuration_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 35000
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-08"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Цена обновлена"
                    },
                    "404": {
                        "description": "Цена не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Prices"
                ],
                "summary": "Удалить цену",
                "operationId": "affc9c47028dfff617ab6f9da7280b26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID цены",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Цена удалена"
                    },
                    "404": {
                        "description": "Цена не найдена"
                    }
                }
            }
        }
    },
...
}